pyproject.toml
[tool.poetry]
name = "condor-metrics"
version = "0.1.0"
description = ""
authors = ["Boris <bio-boris@github.com>"]
readme = "README2.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.1"
uvicorn = "^0.23.2"
htcondor = "^23.1.0"
pysocks = "^1.7.1"

# Dev dependencies
[tool.poetry.dev-dependencies]
pytest = "^6.2.5"
pytest-cov = "^3.0.0"
black = "^21.9b0"
flake8 = "^4.0.1"
flake8-black = "^0.2.2"
flake8-bugbear = "^21.9.2"
flake8-docstrings = "^1.6.0"
flake8-import-order = "^0.18.1"
flake8-annotations = "^2.6.2"
httpx = "^0.19.0"
mypy = "^0.910"
mypy-extensions = "^0.4.3"




[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"